Starting val2y()

+++ src/HeatMapGenerator/Line.java
@@ -1,6 +1,6 @@
 package HeatMapGenerator;
 
 public class Line {
-	public int m;
-	public int b;
+	public double m;
+	public double b;
 }
